Fixed: AuthorFolderAsRootFolderValidator
(cherry picked from commit 0ce81e1ab69d43fde382cc4ae22cd46fe626dea7)

# Conflicts:
#	src/Sonarr.Api.V3/Series/SeriesFolderAsRootFolderValidator.cs
